rearrange some functions and a new expander header
Add comments to more functions and also create a separate expander header file for better overview.
Also some minor changes regarding the source code files (arranging functions).
