+++ include/environment.h
@@ -3,36 +3,43 @@
 
 # include "defines.h"
 
-/* `<summary>`:
- Initializes the main struct from minishell.
- `<parameter>`:
- `main`: main struct containing all other variables;
- `env`: array of strings containing the env variables;
- `<returns>`:
- Returns `TRUE` on success and `FALSE` on fail. */
+/* `<SUMMARY>`:
+ * Initializes the main struct from minishell.
+ * `<PARAM>`:
+ * `main`: main struct containing all other variables;
+ * `env`: array of strings containing the env variables;
+ * `<RETURN>`:
+ * Returns `TRUE` on success and `FALSE` on fail. */
 t_bool	init_shell(t_main *main, char **env);
 
-/* `<summary>`:
- Creates a default env when no envs are set.
- `<parameter>`:
- `main`: main struct containing all other variables;
- `env`: array of strings containing the env variables;
- `<returns>`:
- `TRUE` on success and `FALSE` on fail. */
-t_bool	create_own_environment(t_main *main);
-
-/* `<summary>`:
-Checks if the variable `name` is a variable in the
-environment array of strings.
- `<parameter>`:
- `env`: string member of the array of strings of the environment;
- `name`: variable name to check against;
- `<returns>`:
- `TRUE` on success and `FALSE` on fail. */
+/* `<SUMMARY>`:
+ * Checks if the variable `name` is a variable in the
+ * environment array of strings.
+ * `<PARAM>`:
+ * `env`: string member of the array of strings of the environment;
+ * `name`: variable name to check against;
+ * `<RETURN>`:
+ * `TRUE` on success and `FALSE` on fail. */
 t_bool	check_env_variable(char *env, char *name);
 
+/* `<SUMMARY>`:
+ * Gets the variable `search` from the environment variables.
+ * `<PARAM>`:
+ * `env`: string member of the array of strings of the environment;
+ * `search`: variable name to search for;
+ * `<RETURN>`:
+ * Freeable string of the found string or `NULL`. */
 char	*get_env(char **env, char *search);
 
+/* `<SUMMARY>`:
+ * Updates the environment array of strings in the main sh struct
+ * with the given `var`. If found, it will replace it, else it will
+ * append the string to the array..
+ * `<PARAM>`:
+ * `sh`: main struct containing also the environment;
+ * `var`: variable to update the environment with;
+ * `<RETURN>`:
+ * Updated environment array of strings. */
 char	**update_env_variables(t_main *sh, char *var);
 
 #endif
