+++ include/minishell.h
@@ -4,6 +4,7 @@
 # include <limits.h>
 # include <errno.h>
 # include "defines.h"
+# include "expander.h"
 # include "environment.h"
 # include "signal_handler.h"
 # include "lexer.h"
