+++ src/expander/expander_utils.c
@@ -1,5 +1,30 @@
 #include "minishell.h"
 
+/* `<SUMMARY>`:
+ * .Helper function for creating the key list for the expansion.
+ *  Copies the key and advances the char pointer.
+ * `<PARAM>`:
+ * `input`: Pointer to the input string;
+ * `size`: size of the key;
+ * `<RETURN>`:
+ * Returns a freeable copy of the key (`$` is being removed). */
+static char	*advance_and_copy(char **input, int size)
+{
+	char	*key;
+
+	key = NULL;
+	key = ft_strndup(*input, size);
+	*input += size;
+	return (key);
+}
+
+/* `<SUMMARY>`:
+ * .Function to extract the key from a given string.
+ * Checks for `$` and decides accordingly.
+ * `<PARAM>`:
+ * `input`: Pointer to the input string;
+ * `<RETURN>`:
+ * Returns a freeable copy of the key (`$` is being removed). */
 static char	*get_key(char **input)
 {
 	t_bool	check;
@@ -40,16 +65,6 @@ t_list	*create_keylist(char *input)
 		ft_lstadd_back(&start, ft_lstnew(key));
 	}
 	return (start);
-}
-
-char	*advance_and_copy(char **input, int size)
-{
-	char	*key;
-
-	key = NULL;
-	key = ft_strndup(*input, size);
-	*input += size;
-	return (key);
 }
 
 char	*expand_special(char *input)
