+++ src/main.c
@@ -1,12 +1,5 @@
 #include "minishell.h"
 
-// main loop
-// signal_handler
-// readline
-// lexer
-// parser -> expander
-// executer
-// readline
 #include <math.h>
 
 char* types[] = {
@@ -103,65 +96,67 @@ void	print_AST(t_cmd *cmd)
 	}
 }
 
-// int	main(int argc, char *argv[], char *envp[])
-// {
-// 	t_main	sh;
-// 	t_cmd	*ast;
-// 	char	*str;
-// 
-// 	sh = (t_main){};
-// 	init_shell(&sh, envp);
-// 	str = readline(" > ");
-// 	sh.lexer = ft_lexer(str);
-// 	ast = parse_command(&sh.lexer.token_list, &sh);
-// 	print_AST(ast);
-// }
-
 int	main(int argc, char *argv[], char *envp[])
 {
-	int			len;
-	t_main		sh;
-	char		*str;
+	t_main	sh;
 	t_cmd	*ast;
+	char	*str;
 
 	sh = (t_main){};
-	set_exit_status(EXIT_SUCCESS);
+	ast = NULL;
+	str = NULL;
 	init_shell(&sh, envp);
-	// main.lexer = ft_lexer("cat < test.txt | grep hello | wc > out.log");
 	str = readline(" > ");
-	len = ft_strlen(str);
 	sh.lexer = ft_lexer(str);
-	// main.lexer = ft_lexer("'test' < \"test\" | 'hello'");
-	ft_print_token_list(&sh.lexer);
-	// main.cmd = ft_parser(&main.lexer);
 	ast = parse_command(&sh.lexer.token_list, &sh);
-	// printf("Input: ");
-	// for(int i = 0; i <= len; i++) {
-	// 	if (str[i] == '\0') {
-	// 		printf("\\0"); // print \0 for null character
-	// 	} else {
-	// 		printf("%c", str[i]);
-	// 	}
-	// }
-	// printf("\n");
 	print_AST(ast);
-	executor(ast);
-	// // CLEANUP
-	clean_ast(ast);
-	ft_dlstclear(&sh.lexer.token_list, &free);
-	ft_arrfree(sh.env);
-	ft_arrfree(sh.bin_path);
-	free(str);
-	// ft_print_ast(main.cmd, "START");
-	// ft_arrprint((const char **)main.env);
-	// printf("%lu\n\n", ft_arrlen((const char **)main.env));
-	// env(main.env);
-	// ft_unset(main.env, "USER");
-	// env(main.env);
-	// // printf("%lu\n\n", ft_arrlen((const char **)main.env));
-	// // ft_arrprint((const char **)main.env);
-	// ft_arrfree(main.bin_path);
-	// ft_arrfree(main.env);
-
-	return (EXIT_SUCCESS);
 }
+
+// int	main(int argc, char *argv[], char *envp[])
+// {
+// 	int			len;
+// 	t_main		sh;
+// 	char		*str;
+// 	t_cmd	*ast;
+// 
+// 	sh = (t_main){};
+// 	set_exit_status(EXIT_SUCCESS);
+// 	init_shell(&sh, envp);
+// 	// main.lexer = ft_lexer("cat < test.txt | grep hello | wc > out.log");
+// 	str = readline(" > ");
+// 	len = ft_strlen(str);
+// 	sh.lexer = ft_lexer(str);
+// 	// main.lexer = ft_lexer("'test' < \"test\" | 'hello'");
+// 	ft_print_token_list(&sh.lexer);
+// 	// main.cmd = ft_parser(&main.lexer);
+// 	ast = parse_command(&sh.lexer.token_list, &sh);
+// 	// printf("Input: ");
+// 	// for(int i = 0; i <= len; i++) {
+// 	// 	if (str[i] == '\0') {
+// 	// 		printf("\\0"); // print \0 for null character
+// 	// 	} else {
+// 	// 		printf("%c", str[i]);
+// 	// 	}
+// 	// }
+// 	// printf("\n");
+// 	print_AST(ast);
+// 	executor(ast);
+// 	// // CLEANUP
+// 	clean_ast(ast);
+// 	ft_dlstclear(&sh.lexer.token_list, &free);
+// 	ft_arrfree(sh.env);
+// 	ft_arrfree(sh.bin_path);
+// 	free(str);
+// 	// ft_print_ast(main.cmd, "START");
+// 	// ft_arrprint((const char **)main.env);
+// 	// printf("%lu\n\n", ft_arrlen((const char **)main.env));
+// 	// env(main.env);
+// 	// ft_unset(main.env, "USER");
+// 	// env(main.env);
+// 	// // printf("%lu\n\n", ft_arrlen((const char **)main.env));
+// 	// // ft_arrprint((const char **)main.env);
+// 	// ft_arrfree(main.bin_path);
+// 	// ft_arrfree(main.env);
+// 
+// 	return (EXIT_SUCCESS);
+// }
